[API] add write-read operation and add length to read
The IPC Array used in the to read operation does not transmit the
client selected length of the array to the server. Therefore, the
array cannot be used to transmit the length of the read request.
The new parameter fixes this.

The new write_read interface allows the client to do a combined
write-read transaction on the i2c device. The usual case is to first
write the register to read from in the second transaction.

Change-Id: I2a5f5002bb43c73a0c3307023d1615461a96a1c8